+++ js/data.js
@@ -1,56 +1,6 @@
-import {getPositiveInteger, getPositiveFloat, getRandomItemsArray, MIN_RANDOM_VALUE} from './utils.js';
-
-const ADVERTISEMENT_TITLES = [
-  'Самая лучшая квартира',
-  'Такой квартиры больше нет',
-  'Выбирайте квартиру мечты всех котят',
-  'Сдаем только кошатникам',
-  'Квартира для кошечек',
-];
-const ADVERTISEMENT_DESCRIPTIONS = [
-  'Описание 1',
-  'Описание 2',
-  'Описание 3',
-  'Описание 4',
-  'Описание 5',
-];
-const PLACE_TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel',
-];
-const CHECKIN_CHECKOUT_TIMES = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-const FEATURES_TYPES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-];
-const PHOTOS_LINKS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
-];
-
-const LAT_MIN_VALUE = 35.65000;
-const LAT_MAX_VALUE = 35.70000;
-const LNG_MIN_VALUE = 139.70000;
-const LNG_MAX_VALUE = 139.80000;
-const LOCATION_VALUE_ACCURACY = 5;
-
-const MAX_AVATAR_INDEX = 10;
-const MAX_PRICE = 100;
-const MAX_ROOMS = 6;
-const MAX_GUESTS = 8;
-
+import {getPositiveInteger, getPositiveFloat, getRandomItemsArray} from './utils.js';
+import {MIN_RANDOM_VALUE} from './constants/general.js';
+import {ADVERTISEMENT_TITLES,ADVERTISEMENT_DESCRIPTIONS,PLACE_TYPES,CHECKIN_CHECKOUT_TIMES,FEATURES_TYPES,PHOTOS_LINKS,LATITUDE,LONGTITUDE,LOCATION_VALUE_ACCURACY,MAX_AVATAR_INDEX,MAX_PRICE,MAX_ROOMS,MAX_GUESTS} from './mocks.js';
 
 const createAuthor = () => {
   const avatarIndex =  getPositiveInteger(MIN_RANDOM_VALUE,MAX_AVATAR_INDEX);
@@ -61,8 +11,8 @@ const createAuthor = () => {
 };
 
 const createLocation = () => ({
-  lat: getPositiveFloat(LAT_MIN_VALUE, LAT_MAX_VALUE, LOCATION_VALUE_ACCURACY),
-  lng: getPositiveFloat(LNG_MIN_VALUE, LNG_MAX_VALUE, LOCATION_VALUE_ACCURACY),
+  lat: getPositiveFloat(LATITUDE.min, LONGTITUDE.max, LOCATION_VALUE_ACCURACY),
+  lng: getPositiveFloat(LONGTITUDE.min, LONGTITUDE.max, LOCATION_VALUE_ACCURACY),
 });
 
 const createOffer = (location) => {
