+++ js/mocks.js
@@ -1,0 +1,51 @@
+
+const ADVERTISEMENT_TITLES = [
+  'Самая лучшая квартира',
+  'Такой квартиры больше нет',
+  'Выбирайте квартиру мечты всех котят',
+  'Сдаем только кошатникам',
+  'Квартира для кошечек',
+];
+const ADVERTISEMENT_DESCRIPTIONS = [
+  'Описание 1',
+  'Описание 2',
+  'Описание 3',
+  'Описание 4',
+  'Описание 5',
+];
+const PLACE_TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+const CHECKIN_CHECKOUT_TIMES = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const FEATURES_TYPES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const PHOTOS_LINKS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const LATITUDE = {min:35.65, max:35.70};
+const LONGTITUDE = {min:139.7, max:139.8};
+const LOCATION_VALUE_ACCURACY = 5;
+
+const MAX_AVATAR_INDEX = 10;
+const MAX_PRICE = 100;
+const MAX_ROOMS = 6;
+const MAX_GUESTS = 8;
+
+export {ADVERTISEMENT_TITLES,ADVERTISEMENT_DESCRIPTIONS,PLACE_TYPES,CHECKIN_CHECKOUT_TIMES,FEATURES_TYPES,PHOTOS_LINKS,LATITUDE,LONGTITUDE,LOCATION_VALUE_ACCURACY,MAX_AVATAR_INDEX,MAX_PRICE,MAX_ROOMS,MAX_GUESTS};
